fix Bug #70651, set thread org id for GroupThread, make sure mv created with right org id.
